fix(toolkit): use correct hook for fetching auth code (#457)
* fix(toolkit): use correct hook for fetching auth code

* fix(toolkit): use correct hook for fetching auth code

* fix(toolkit): fix next issue
