+++ src/interfaces/coral_web/src/utils/queryString.ts
@@ -5,7 +5,7 @@
  *    : router.query.[param];
  */
 
-export const getQueryString = (param?: string | string[]) => {
+export const getQueryString = (param?: string | string[] | null) => {
   if (!param) return;
   if (Array.isArray(param)) {
     return param[0] as string;
